d/shuffle-binaries: Switch to POSIX shell, instead of Bash
While bash is great for an interactive shell, sh is more lightweight and
portable and is more suited for scripts that need to be executed fast
and do not require user-friendly interaction with an end user.

Drop pipefail, since it has less value now that the inner loop of the
script has been simplified.

Signed-off-by: Elliott Mitchell <ehem+debian@m5p.com>
Acked-by: Hans van Kranenburg <hans@knorrie.org>

[Hans van Kranenburg]
Reduce usage of dramatic language in commit message