xen/arm: traps: Don't panic when receiving an unknown debug trap
Even if debug trap are only meant for debugging purpose, it is quite
harsh to crash Xen if one of the trap sent by the guest is not handled.

So switch from a panic() to a printk().

Signed-off-by: Julien Grall <jgrall@amazon.com>
Reviewed-by: Stefano Stabellini <sstabellini@kernel.org>
(cherry picked from commit 957708c2d1ae25d7375abd5e5e70c3043d64f1f1)

Gbp-Pq: Name 0035-xen-arm-traps-Don-t-panic-when-receiving-an-unknown-.patch